requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
neo4j>=5.20.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0